requirements.txt
asgiref==3.8.1
Django
django-extensions==3.2.3
livereload==2.7.0
numpy
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
whitenoise==6.7.0
djangorestframework
onesignal_sdk
opencv-python
djangorestframework-simplejwt
ultralytics
